"Duck" organize container/presentation components and related stores. (#4)
