+++ src/App.tsx
@@ -1,14 +1,14 @@
 import * as React from 'react';
 import { Route, Switch } from 'react-router-dom';
-import DepartmentContainer from './components/DepartmentContainer';
-import DepartmentsContainer from './components/DepartmentsContainer';
-import Home from './components/Home';
+import DepartmentContainer from './components/Department/Container';
+import DepartmentsContainer from './components/Departments/Container';
+import Home from './components/Home/Home';
 import Page from './components/layout/Page';
-import ProfileContainer from './components/ProfileContainer';
-import Signin from './components/Signin';
-import SimpleSearchContainer from "./components/SimpleSearchContainer";
-import UnitContainer from "./components/UnitContainer";
-import UnitsContainer from "./components/UnitsContainer";
+import ProfileContainer from './components/Profile/Container';
+import SearchContainer from "./components/Search/Container";
+import Signin from './components/SignIn/Component';
+import UnitContainer from "./components/Unit/Container";
+import UnitsContainer from "./components/Units/Container";
 
 const App : React.SFC = () => (
   <Page>
@@ -17,7 +17,7 @@ const App : React.SFC = () => (
       <Route path="/signin" component={Signin} />
       <Route path="/profiles/:id" component={ProfileContainer} /> 
       <Route path="/me" component={ProfileContainer} /> 
-      <Route path="/search" component={SimpleSearchContainer} /> 
+      <Route path="/search" component={SearchContainer} /> 
       <Route path="/departments/:id" component={DepartmentContainer} /> 
       <Route path="/departments" component={DepartmentsContainer} /> 
       <Route path="/units/:id" component={UnitContainer} />
