+++ src/components/Department/store.ts
@@ -1,5 +1,9 @@
-import { IApiState } from './common'
-import { IEntity } from './profile';
+import { Reducer  } from 'redux'
+import { all, fork, select, takeEvery } from 'redux-saga/effects'
+import { action } from 'typesafe-actions'
+import { httpGet } from '../effects'
+import { FetchErrorReducer, FetchRequestReducer, FetchSuccessReducer } from '../types'
+import { IApiState, IApplicationState, IEntity } from '../types'
 
 //#region TYPES
 export const enum DepartmentActionTypes {
@@ -23,29 +27,23 @@ export interface IState extends IApiState<IDepartmentRequest, IDepartmentProfile
 }
 //#endregion
 
-//#region ACTIONS
-import { action } from 'typesafe-actions'
 
-const fetchRequest = (request: IDepartmentRequest) => action(DepartmentActionTypes.DEPARTMENT_FETCH_REQUEST, request)
-const fetchSuccess = (data: IDepartmentProfile) => action(DepartmentActionTypes.DEPARTMENT_FETCH_SUCCESS, data)
-const fetchError = (error: string) => action(DepartmentActionTypes.DEPARTMENT_FETCH_ERROR, error)
+//#region ACTIONS
+export const fetchRequest = (request: IDepartmentRequest) => action(DepartmentActionTypes.DEPARTMENT_FETCH_REQUEST, request)
+export const fetchSuccess = (data: IDepartmentProfile) => action(DepartmentActionTypes.DEPARTMENT_FETCH_SUCCESS, data)
+export const fetchError = (error: string) => action(DepartmentActionTypes.DEPARTMENT_FETCH_ERROR, error)
 //#endregion
 
-//#region REDUCER
-import { Reducer } from 'redux'
-import { FetchErrorReducer, FetchRequestReducer, FetchSuccessReducer } from './common'
 
-// Type-safe initialState!
-const initialState: IState = {
+//#region REDUCER
+export const initialState: IState = {
     data: undefined,
     error: undefined,
     loading: false,
     request: undefined,
 }
 
-// Thanks to Redux 4's much simpler typings, we can take away a lot of typings on the reducer side,
-// everything will remain type-safe.
-const reducer: Reducer<IState> = (state = initialState, act) => {
+export const reducer: Reducer<IState> = (state = initialState, act) => {
   switch (act.type) {
     case DepartmentActionTypes.DEPARTMENT_FETCH_REQUEST: return FetchRequestReducer(state, act)
     case DepartmentActionTypes.DEPARTMENT_FETCH_SUCCESS: return FetchSuccessReducer(state, act)
@@ -57,12 +55,8 @@ const reducer: Reducer<IState> = (state = initialState, act) => {
 
 
 //#region SAGA
-import { all, fork, select, takeEvery } from 'redux-saga/effects'
-import { httpGet } from './effects'
-import { IApplicationState } from './index';
-
 function* handleFetch() {
-    const state = (yield select<IApplicationState>((s) => s.org.request)) as IDepartmentRequest
+    const state = (yield select<IApplicationState>((s) => s.department.request)) as IDepartmentRequest
     const path = `/departments/${state.id}`
     yield httpGet<IDepartmentProfile>(path, fetchSuccess, fetchError)
 }
@@ -74,19 +68,7 @@ function* watchDepartmentFetch() {
 }
 
 // We can also use `fork()` here to split our saga into multiple watchers.
-function* saga() {
+export function* saga() {
   yield all([fork(watchDepartmentFetch)])
 }
 //#endregion
-
-
-// Instead of using default export, we use named exports. That way we can group these exports
-// inside the `index.js` folder.
-export { 
-  fetchRequest,
-  fetchError,
-  fetchSuccess,
-  reducer,
-  initialState,
-  saga
-}
