+++ src/components/Departments/Presentation.tsx
@@ -1,11 +1,9 @@
 import * as React from 'react'
 import { Panel, Table } from "rivet-react";
-import { IDepartmentList } from "../store/departments";
+import { IDepartmentList } from "./store";
 
-const Departments: React.SFC<IDepartmentList> =
+const Presentation: React.SFC<IDepartmentList> =
     (props) =>{ 
-        
-        console.log("departments", props)
         return (
         <>
             <Panel margin={{ top: "xs" }}>
@@ -23,4 +21,4 @@ const Departments: React.SFC<IDepartmentList> =
             </Panel>
         </>
     )}
-export default Departments
+export default Presentation
