+++ src/components/Departments/store.ts
@@ -1,5 +1,4 @@
-import { IApiState } from './common'
-import { IEntity } from './profile';
+import { IApiState, IEntity } from '../types'
 
 //#region TYPES
 export const enum DepartmentsActionTypes {
@@ -18,17 +17,17 @@ export interface IState extends IApiState<{}, IDepartmentList> {}
 //#region ACTIONS
 import { action } from 'typesafe-actions'
 
-const fetchRequest = () => action(DepartmentsActionTypes.DEPARTMENTS_FETCH_REQUEST)
-const fetchSuccess = (data: IDepartmentList) => action(DepartmentsActionTypes.DEPARTMENTS_FETCH_SUCCESS, data)
-const fetchError = (error: string) => action(DepartmentsActionTypes.DEPARTMENTS_FETCH_ERROR, error)
+export const fetchRequest = () => action(DepartmentsActionTypes.DEPARTMENTS_FETCH_REQUEST)
+export const fetchSuccess = (data: IDepartmentList) => action(DepartmentsActionTypes.DEPARTMENTS_FETCH_SUCCESS, data)
+export const fetchError = (error: string) => action(DepartmentsActionTypes.DEPARTMENTS_FETCH_ERROR, error)
 //#endregion
 
 //#region REDUCER
 import { Reducer } from 'redux'
-import { FetchErrorReducer, FetchRequestReducer, FetchSuccessReducer } from './common'
+import { FetchErrorReducer, FetchRequestReducer, FetchSuccessReducer } from '../types'
 
 // Type-safe initialState!
-const initialState: IState = {
+export const initialState: IState = {
     data: undefined,
     error: undefined,
     loading: false,
@@ -37,7 +36,7 @@ const initialState: IState = {
 
 // Thanks to Redux 4's much simpler typings, we can take away a lot of typings on the reducer side,
 // everything will remain type-safe.
-const reducer: Reducer<IState> = (state = initialState, act) => {
+export const reducer: Reducer<IState> = (state = initialState, act) => {
   switch (act.type) {
     case DepartmentsActionTypes.DEPARTMENTS_FETCH_REQUEST: return FetchRequestReducer(state, act)
     case DepartmentsActionTypes.DEPARTMENTS_FETCH_SUCCESS: return FetchSuccessReducer(state, act)
@@ -50,7 +49,7 @@ const reducer: Reducer<IState> = (state = initialState, act) => {
 
 //#region SAGA
 import { all, fork, takeEvery } from 'redux-saga/effects'
-import { httpGet } from './effects'
+import { httpGet } from '../effects'
 
 function* handleFetch() {
   yield httpGet<IDepartmentList>("/departments", fetchSuccess, fetchError)
@@ -63,19 +62,7 @@ function* watchDepartmentsFetch() {
 }
 
 // We can also use `fork()` here to split our saga into multiple watchers.
-function* saga() {
+export function* saga() {
   yield all([fork(watchDepartmentsFetch)])
 }
 //#endregion
-
-
-// Instead of using default export, we use named exports. That way we can group these exports
-// inside the `index.js` folder.
-export { 
-  fetchRequest,
-  fetchError,
-  fetchSuccess,
-  reducer,
-  initialState,
-  saga
-}
