--- src/components/Home.tsx
@@ -1,11 +1,0 @@
-
-import * as React from 'react';
-import PageTitle from './layout/PageTitle';
-
-const Home: React.SFC = () => (
-  <PageTitle>Home</PageTitle>
-);
-
-Home.displayName = 'Home';
-
-export default Home;
