+++ src/components/Home/Home.tsx
@@ -1,0 +1,11 @@
+
+import * as React from 'react';
+import PageTitle from '../layout/PageTitle';
+
+const Component: React.SFC = () => (
+  <PageTitle>Home</PageTitle>
+);
+
+Component.displayName = 'Home';
+
+export default Component;
