+++ src/components/Search/Container.tsx
@@ -2,10 +2,10 @@ import * as queryString from 'query-string'
 import * as React from 'react';
 import { connect } from 'react-redux';
 import { Dispatch } from 'redux';
-import { IApplicationState } from '../store';
-import * as searchSimple from '../store/searchSimple';
-import PageTitle from './layout/PageTitle';
-import SimpleSearch from './SimpleSearch';
+import PageTitle from '../layout/PageTitle';
+import { IApplicationState } from '../types';
+import Search from './Presentation';
+import { fetchRequest, ISimpleSearchRequest, IState } from './store';
 
 interface ILocationProps {
     search: string;
@@ -15,22 +15,19 @@ interface ISearchProps {
     location: ILocationProps
 }
 
-// We can use `typeof` here to map our dispatch types to the props, like so.
 interface IPropsFromDispatch {
-    searchRequest: typeof searchSimple.fetchRequest
+    searchRequest: typeof fetchRequest
 }
 
-interface ISimpleSearchContainerProps extends searchSimple.IState, ISearchProps, IPropsFromDispatch {
-
-}
-
+interface ISimpleSearchContainerProps 
+    extends IState, ISearchProps, IPropsFromDispatch {}
+    
 const executeSearch = (props: ISimpleSearchContainerProps) => {
     const queryParam = queryString.parse(props.location.search)
     props.searchRequest({ term: queryParam.term })   
 }
 
-// tslint:disable-next-line:max-classes-per-file
-class SimpleSearchContainer extends React.Component<ISimpleSearchContainerProps>{
+class Container extends React.Component<ISimpleSearchContainerProps>{
     
     public componentDidMount() {
         executeSearch(this.props)
@@ -49,7 +46,7 @@ class SimpleSearchContainer extends React.Component<ISimpleSearchContainerProps>
                 { this.props.loading && 
                     <p>Searching...</p>}
                 { this.props.data &&  
-                    <SimpleSearch {...this.props.data} /> }
+                    <Search {...this.props.data} /> }
                 { this.props.error && 
                     <p>{this.props.error}</p> }
             </>
@@ -57,21 +54,14 @@ class SimpleSearchContainer extends React.Component<ISimpleSearchContainerProps>
     }
 }
 
-// Although if necessary, you can always include multiple contexts. Just make sure to
-// separate them from each other to prevent prop conflicts.
-const mapStateToProps = (state: IApplicationState) => ({
-  ...state.searchSimple
-})
+const mapStateToProps = (state: IApplicationState) => 
+    state.searchSimple
   
-// mapDispatchToProps is especially useful for constraining our actions to the connected component.
-// You can access these via `this.props`.
 const mapDispatchToProps = (dispatch: Dispatch) : IPropsFromDispatch => ({
-  searchRequest: (request: searchSimple.ISimpleSearchRequest) => dispatch(searchSimple.fetchRequest(request))
+  searchRequest: (request: ISimpleSearchRequest) => dispatch(fetchRequest(request))
 })
   
-// Now let's connect our component!
-// With redux v4's improved typings, we can finally omit generics here.
 export default connect(
     mapStateToProps,
     mapDispatchToProps
-  )(SimpleSearchContainer)
+  )(Container)
