+++ src/components/Search/store.ts
@@ -1,5 +1,4 @@
-import { IApiState } from './common'
-import { IEntity } from './profile';
+import { IApiState, IApplicationState, IEntity } from '../types'
 
 //#region TYPES
 export const enum SearchActionTypes {
@@ -26,27 +25,24 @@ export interface IState extends IApiState<ISimpleSearchRequest, ISimpleSearchRes
 //#region ACTIONS
 import { action } from 'typesafe-actions'
 
-const submit = () => action(SearchActionTypes.SEARCH_SIMPLE_SUBMIT)
-const fetchRequest = (request: ISimpleSearchRequest) => action(SearchActionTypes.SEARCH_SIMPLE_FETCH_REQUEST, request)
-const fetchSuccess = (data: ISimpleSearchResult) => action(SearchActionTypes.SEARCH_SIMPLE_FETCH_SUCCESS, data)
-const fetchError = (error: string) => action(SearchActionTypes.SEARCH_SIMPLE_FETCH_ERROR, error)
+export const submit = () => action(SearchActionTypes.SEARCH_SIMPLE_SUBMIT)
+export const fetchRequest = (request: ISimpleSearchRequest) => action(SearchActionTypes.SEARCH_SIMPLE_FETCH_REQUEST, request)
+export const fetchSuccess = (data: ISimpleSearchResult) => action(SearchActionTypes.SEARCH_SIMPLE_FETCH_SUCCESS, data)
+export const fetchError = (error: string) => action(SearchActionTypes.SEARCH_SIMPLE_FETCH_ERROR, error)
 //#endregion
 
 //#region REDUCER
 import { Reducer } from 'redux'
-import { FetchErrorReducer, FetchRequestReducer, FetchSuccessReducer } from './common'
+import { FetchErrorReducer, FetchRequestReducer, FetchSuccessReducer } from '../types'
 
-// Type-safe initialState!
-const initialState: IState = {
+export const initialState: IState = {
     data: undefined,
     error: undefined,
     loading: false,
     request: undefined,
 }
 
-// Thanks to Redux 4's much simpler typings, we can take away a lot of typings on the reducer side,
-// everything will remain type-safe.
-const reducer: Reducer<IState> = (state = initialState, act) => {
+export const reducer: Reducer<IState> = (state = initialState, act) => {
   switch (act.type) {
     case SearchActionTypes.SEARCH_SIMPLE_FETCH_REQUEST: return FetchRequestReducer(state, act)
     case SearchActionTypes.SEARCH_SIMPLE_FETCH_SUCCESS: return FetchSuccessReducer(state, act)
@@ -60,9 +56,8 @@ const reducer: Reducer<IState> = (state = initialState, act) => {
 //#region SAGA
 import { push } from 'react-router-redux';
 import { all, fork, put, select, takeEvery,  } from 'redux-saga/effects'
-import { httpGet } from './effects'
-import { IApplicationState } from './index';
-import { IUnitList } from './units';
+import { httpGet } from '../effects'
+import { IUnitList } from '../Units/store';
 
 function* handleFetch() {
     const state = (yield select<IApplicationState>((s) => s.searchSimple.request)) as ISimpleSearchRequest
@@ -77,8 +72,6 @@ function* handleSubmit() {
   }
 }
 
-// This is our watcher function. We use `take*()` functions to watch Redux for a specific action
-// type, and run our saga, for example the `handleFetch()` saga above.
 function* watchSimpleSearchFetch() {
   yield takeEvery(SearchActionTypes.SEARCH_SIMPLE_FETCH_REQUEST, handleFetch)
 }
@@ -87,21 +80,7 @@ function* watchSimpleSearchSubmit() {
   yield takeEvery(SearchActionTypes.SEARCH_SIMPLE_SUBMIT, handleSubmit)
 }
 
-// We can also use `fork()` here to split our saga into multiple watchers.
-function* saga() {
+export function* saga() {
   yield all([fork(watchSimpleSearchFetch), fork(watchSimpleSearchSubmit)])
 }
 //#endregion
-
-
-// Instead of using default export, we use named exports. That way we can group these exports
-// inside the `index.js` folder.
-export { 
-  submit,
-  fetchRequest,
-  fetchError,
-  fetchSuccess,
-  reducer,
-  initialState,
-  saga
-}
