+++ src/components/Unit/Container.tsx
@@ -1,22 +1,22 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
 import { Dispatch } from 'redux';
-import { IApplicationState } from '../store';
-import * as org from '../store/department';
-import Department from './Department';
-import PageTitle from './layout/PageTitle';
+import PageTitle from '../layout/PageTitle';
+import { IApplicationState } from '../types';
+import Unit from './Presentation';
+import * as unit from './store';
 
-interface IProps {
+interface IContainerProps {
     match: any
 }
 
 // We can use `typeof` here to map our dispatch types to the props, like so.
-interface IPropsFromDispatch {
-    fetchRequest: typeof org.fetchRequest
+interface IDispatchProps {
+    fetchRequest: typeof unit.fetchRequest
 }
 
 // tslint:disable-next-line:max-classes-per-file
-class UnitContainer extends React.Component<org.IState & IProps & IPropsFromDispatch>{
+class Container extends React.Component<unit.IState & IContainerProps & IDispatchProps>{
     public componentDidMount() {
         console.log(this.props.match)
         this.props.fetchRequest(this.props.match.params)
@@ -26,13 +26,12 @@ class UnitContainer extends React.Component<org.IState & IProps & IPropsFromDisp
         return (
             <>
                 {/* { this.props.loading && 
-                    <p>Loading department...</p>} */}
+                    <p>Loading unit...</p>} */}
                 { this.props.data &&
                     <>
-                        <PageTitle>{`${this.props.data.department.description} (${this.props.data.department.name})`}</PageTitle>
-                        <Department {...this.props.data} />
-                    </> 
-                }
+                        <PageTitle>{this.props.data.unit.name}</PageTitle>
+                        <Unit {...this.props.data} />
+                    </> }
                 { this.props.error && 
                     <p>{this.props.error}</p> }
             </>
@@ -43,13 +42,13 @@ class UnitContainer extends React.Component<org.IState & IProps & IPropsFromDisp
 // Although if necessary, you can always include multiple contexts. Just make sure to
 // separate them from each other to prevent prop conflicts.
 const mapStateToProps = (state: IApplicationState) => ({
-  ...state.org
+  ...state.unit
 })
   
 // mapDispatchToProps is especially useful for constraining our actions to the connected component.
 // You can access these via `this.props`.
-const mapDispatchToProps = (dispatch: Dispatch) : IPropsFromDispatch => ({
-  fetchRequest: (request: org.IDepartmentRequest) => dispatch(org.fetchRequest(request))
+const mapDispatchToProps = (dispatch: Dispatch) : IDispatchProps => ({
+  fetchRequest: (request: unit.IUnitRequest) => dispatch(unit.fetchRequest(request))
 })
   
 // Now let's connect our component!
@@ -57,4 +56,4 @@ const mapDispatchToProps = (dispatch: Dispatch) : IPropsFromDispatch => ({
 export default connect(
     mapStateToProps,
     mapDispatchToProps
-  )(UnitContainer)
+  )(Container)
