+++ src/components/layout/Page.tsx
@@ -2,9 +2,9 @@ import * as React from 'react';
 import { connect } from 'react-redux'
 import { Dispatch } from 'redux'
 import { Container, Footer, Header, HeaderIdentity, HeaderNavigation } from 'rivet-react';
-import { IApplicationState  } from '../../store'
-import * as Auth from '../../store/auth'
-import * as Search from '../../store/searchSimple'
+import * as Search from '../Search/store'
+import * as Auth from '../SignIn/store'
+import { IApplicationState  } from '../types'
 import SearchForm from './SearchForm';
 
 export interface IPageProps {
